src/requirements.txt
discord.py==2.2.2
fastapi==0.95.1
google-api-python-client==2.85.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
ics==0.7.2
python-dateutil==2.8.2
sqlmodel==0.0.8
uvicorn[standard]==0.21.1
